Created Sublcass folders and usage notes for Example Enemy subclass EnemyTest01.java
